added schedule and running time
Now you can define some value for time:
- `start_time_seconds`
- `start_time_minutes`
- `start_time_hours`
- `start_time_days_of_week`
- `running_timeN`

With this values, you will be able to define not only when do you want to run the sprinklers, but for how long (`using `running_timeN` where `N` is the index for the defined variable you want, if you need multiple run times. 

This way the sprinklers will be more robust and behave accordingly, not depending on HomeAssistant for schedules. 

I've also added a failsafe at the end, where all zones are turned off no matter what.